pyproject.toml
[tool.poetry]
name = "music-metadata-manager"
version = "0.1.0"
description = "A flexible music metadata management system"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "metadata_manager", from = "src" }
]

[tool.poetry.scripts]
metadata-manager = "metadata_manager.cli:main"
metadata-manager-gui = "metadata_manager.gui:main"

[tool.poetry.dependencies]
python = ">3.8,<4.0"
mutagen = "^1.45.1"
musicbrainzngs = "^0.7.1"
rich = "^13.9.4"
requests = "^2.31.0"
Pillow = "^10.0.0"  # Simplificar Pillow
spotipy = "^2.23.0"
ytmusicapi = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
black = "^22.0.0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
